Packages Page
Made it look nice. May as well keep it in. Also added a text file to it to keep track of who worked on what. Feel free to add to that section or we can hand in a separate file with that info on it. We may do that anyway to add workshop four.